Added unit conversion
