+++ frontend/src/utils/conversion/index.ts
@@ -1,0 +1,13 @@
+/**
+ * Unit conversion utilities - barrel export.
+ * Re-exports all conversion functionality for convenient importing.
+ */
+
+// Core conversion utilities
+export * from './unitConversion';
+
+// Time step data conversion
+export * from './timeStepDataConversion';
+
+// Constants conversion
+export * from './constantsConversion';
